chore(dependnencies): upgrade stencil
upgrade stencil to v2.8.0. Versions v2.7.0 of Stencil and up work
on Macs with M1 processors. Specifically, v2.6.0 requires an older
version of puppeteer that does not support ARM architector and
darwin simultanesously, so we must upgrade puppeteer to v10 as well.

by upgrading puppeteer to v10, we can safely remove @types/puppeteer,
since this verison of puppeteer includes it's own typings